appli_typannot/lib/backend/requirements.txt
#SERVEUR WEB
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

#BASE DE DONNÉES
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
pymysql>=1.1.0

#VALIDATION DES DONNÉES
pydantic>=2.0.0
email-validator>=2.1.0

#SÉCURITÉ & AUTHENTIFICATION
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

#CONFIGURATION & ENVIRONNEMENT
python-dotenv>=1.0.0


opencv-python==4.9.0.80
numpy==1.26.4
matplotlib==3.8.2
pandas==2.2.0
dnspython==2.8.0
scp==0.15.0
xlsxwriter==3.2.9
paramiko==4.0.0

sshtunnel
openpyxl